Change to MIT license
This change is done after considering the larger open-source ecosystem of
libraries and applications being developed for Biblical computing. It
has become apparent that most of them are using non-copyleft
licenses. Having USFM-Grammar under the GPL-* license, therefore
may impose on those considering to use this library, a larger decision than is intended.

In our desire to stimulate development and encourage use of
open-source software for Biblical computing, we have decided to go
ahead with using the more permissive MIT license for
USFM-Grammar. Ultimately, our desire is to see His kingdom come.